requirements.txt
setuptools==67.6.1
click==8.1.3
uvloop==0.17.0
coverage==7.2.2
discord.py==2.2.2
gamepack @ git+https://github.com/x4dr/GamePack.git@master#egg=gamepack
bleach==6.0.0
python-dateutil==2.8.2
aiohttp==3.8.4
lavaplayer==1.0.7a0
typing==3.7.4.3
matplotlib==3.7.1
PyNaCl==1.5.0
